move FormModal to shared folder
